+++ src/Scene.cc
@@ -7,37 +7,37 @@ Scene::Scene(std::string file_name)
 
 void Scene::compute_image()
 {
-    cv::Mat3i image = cv::Mat3i(camera_->width_, camera_->height_, cv::Vec3i{0,0,0});
+    cv::Mat3d image = cv::Mat3d(camera_->width_, camera_->height_, cv::Vec3d{0,0,0});
     for (int y = 0; y < camera_->height_; ++y)
         for (int x = 0; x < camera_->width_; ++x)
         {
-            float pX = (2 * ((x + 0.5) / camera_->width_) - 1) * camera_->angle_ * camera_->aspectRatio_;
-            float pY = (1 - 2 * (y + 0.5) / camera_->height_) * camera_->angle_;
+            double pX = (2 * ((x + 0.5) / camera_->width_) - 1) * camera_->angle_ * camera_->aspectRatio_;
+            double pY = (1 - 2 * (y + 0.5) / camera_->height_) * camera_->angle_;
 
             // - rayOrigin in the formula but here its {0, 0, 0} ...
-            cv::Vec3i rayDirection = cv::Vec3i{static_cast<int>(pX), static_cast<int>(pY), -1};
+            cv::Vec3d rayDirection = cv::Vec3d{static_cast<int>(pX), static_cast<int>(pY), -1};
 
             rayDirection = applyTransform(rayDirection, camera_->transformView_);
             rayDirection = cv::normalize(rayDirection);
 
-            image.at<cv::Vec3i>(x, y) = send_ray(cv::Vec3i{0, 0, 0}, rayDirection, 0);
+            image.at<cv::Vec3d>(x, y) = send_ray(cv::Vec3d{0, 0, 0}, rayDirection, 0);
         }
     //show_image(image);
     //save_image("yolo", image);
 }
 
-cv::Vec3i Scene::send_ray(cv::Vec3i rayOrigin, cv::Vec3i rayDirection, int recursion)
+cv::Vec3d Scene::send_ray(cv::Vec3d rayOrigin, cv::Vec3d rayDirection, int recursion)
 {
     std::cout << "FIXME" << std::endl;
-    return cv::Vec3i{0, 0, 0};
+    return cv::Vec3d{0, 0, 0};
 }
 
-cv::Vec3i Scene::applyTransform(cv::Vec3i input, cv::Mat t)
+cv::Vec3d Scene::applyTransform(cv::Vec3d input, cv::Mat t)
 {
-    cv::Vec3i result;
-    result.val[0] = input.dot(cv::Vec3i{t.at<int>(0, 0), t.at<int>(0, 1), t.at<int>(0, 2)});
-    result.val[1] = input.dot(cv::Vec3i{t.at<int>(1, 0), t.at<int>(1, 1), t.at<int>(1, 2)});
-    result.val[2] = input.dot(cv::Vec3i{t.at<int>(2, 0), t.at<int>(2, 1), t.at<int>(2, 2)});
+    cv::Vec3d result;
+    result.val[0] = input.dot(cv::Vec3d{t.at<int>(0, 0), t.at<int>(0, 1), t.at<int>(0, 2)});
+    result.val[1] = input.dot(cv::Vec3d{t.at<int>(1, 0), t.at<int>(1, 1), t.at<int>(1, 2)});
+    result.val[2] = input.dot(cv::Vec3d{t.at<int>(2, 0), t.at<int>(2, 1), t.at<int>(2, 2)});
     return result;
 }
 
@@ -68,9 +68,9 @@ void Scene::load_scene(std::string file_name)
         iss >> word;
         if (word.compare("Camera") == 0)
         {
-            cv::Vec3i e;
-            cv::Vec3i lk;
-            cv::Vec3i u;
+            cv::Vec3d e;
+            cv::Vec3d lk;
+            cv::Vec3d u;
             int fov;
             int w;
             int h;
@@ -80,16 +80,16 @@ void Scene::load_scene(std::string file_name)
         }
         else if (word.compare("Light") == 0)
         {
-            cv::Vec3i pos;
-            cv::Vec3i col;
+            cv::Vec3d pos;
+            cv::Vec3d col;
             iss >> pos.val[0] >> pos.val[1] >> pos.val[2] >> col.val[0] >> col.val[1] >> col.val[2];
             lights_.push_back(Light(pos, col));
         }
         else if (word.compare("Sphere") == 0)
         {
-            cv::Vec3i c;
+            cv::Vec3d c;
             double r;
-            cv::Vec3i col;
+            cv::Vec3d col;
             double alpha;
             shapes::ReflectionType ref;
             iss >> c.val[0] >> c.val[1] >> c.val[2] >> r >> col.val[0] >> col.val[1] >> col.val[2] >> alpha;
@@ -102,10 +102,10 @@ void Scene::load_scene(std::string file_name)
         }
         else if (word.compare("Triangle") == 0)
         {
-            cv::Vec3i p1;
-            cv::Vec3i p2;
-            cv::Vec3i p3;
-            cv::Vec3i col;
+            cv::Vec3d p1;
+            cv::Vec3d p2;
+            cv::Vec3d p3;
+            cv::Vec3d col;
             double alpha;
             shapes::ReflectionType ref;
             iss >> p1.val[0] >> p1.val[1] >> p1.val[2] >> p2.val[0] >> p2.val[1] >> p2.val[2];
