+++ src/Scene.hh
@@ -18,12 +18,12 @@ public:
     Scene(std::string file_name);
 
     void compute_image();
-    cv::Vec3i send_ray(cv::Vec3i rayOrigin, cv::Vec3i rayDirection, int recursion);
+    cv::Vec3d send_ray(cv::Vec3d rayOrigin, cv::Vec3d rayDirection, int recursion);
 
     template <class T>
-    std::pair<T, double> find_intersection(T shape, cv::Vec3i rayOrigin, cv::Vec3i rayDirection);
+    std::pair<T, double> find_intersection(T shape, cv::Vec3d rayOrigin, cv::Vec3d rayDirection);
 
-    cv::Vec3i applyTransform(cv::Vec3i input, cv::Mat t);
+    cv::Vec3d applyTransform(cv::Vec3d input, cv::Mat t);
 
     void show_image(cv::Mat image);
     void save_image(std::string filename, cv::Mat image);
@@ -38,7 +38,7 @@ public:
 };
 
 template <class T>
-std::pair<T, double> Scene::find_intersection(T shape, cv::Vec3i rayOrigin, cv::Vec3i rayDirection)
+std::pair<T, double> Scene::find_intersection(T shape, cv::Vec3d rayOrigin, cv::Vec3d rayDirection)
 {
     std::cout << "lol" << std::endl;
 }
